DFE-837 - rearranged admin dashboard page and its components to sit in single folder
