+++ src/explore-education-statistics-admin/src/pages/prototypes/PrototypePublicationPageAssignMethodology.tsx
@@ -5,8 +5,8 @@ import {
   FormSelect,
   FormRadioGroup,
 } from '@common/components/form';
+import Link from '@admin/components/Link';
 import PrototypePage from './components/PrototypePage';
-import Link from '../../components/Link';
 
 const PublicationPage = () => {
   const [value, setValue] = useState('academic-year');
