+++ src/explore-education-statistics-admin/src/pages/prototypes/PrototypePublicationPageNewAbsenceSchedule.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Link from '../../components/Link';
+import Link from '@admin/components/Link';
 import PrototypeAdminNavigation from './components/PrototypeAdminNavigation';
 import PrototypePage from './components/PrototypePage';
