+++ src/explore-education-statistics-admin/src/pages/prototypes/PrototypePublicationPageNewAbsenceViewTables.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { FormSelect, FormFieldset } from '@common/components/form';
+import Link from '@admin/components/Link';
 import PrototypeAdminExampleTables from './components/PrototypeAdminExampleTables';
 import PrototypeAdminNavigation from './components/PrototypeAdminNavigation';
 import PrototypePage from './components/PrototypePage';
-import Link from '../../components/Link';
 
 import Data from './PrototypeData';
