+++ src/explore-education-statistics-admin/src/pages/prototypes/components/PrototypePublicationPageConfig.tsx
@@ -6,7 +6,7 @@ import {
   FormSelect,
   FormRadioGroup,
 } from '@common/components/form';
-import Link from '../../../components/Link';
+import Link from '@admin/components/Link';
 
 interface Props {
   sectionId?: string;
