+++ src/explore-education-statistics-admin/src/pages/prototypes/components/PrototypeReleasePageConfig.tsx
@@ -5,7 +5,7 @@ import {
   FormTextInput,
   FormRadioGroup,
 } from '@common/components/form';
-import Link from '../../../components/Link';
+import Link from '@admin/components/Link';
 
 interface Props {
   sectionId?: string;
