+++ src/explore-education-statistics-common/src/components/SummaryList.tsx
@@ -6,13 +6,15 @@ import React, { Children, ReactNode } from 'react';
 interface Props {
   children: ReactNode;
   noBorder?: boolean;
+  additionalClassName?: string;
 }
 
-const SummaryList = ({ children, noBorder }: Props) => {
+const SummaryList = ({ children, noBorder, additionalClassName }: Props) => {
   return (
     <dl
       className={classNames('govuk-summary-list', {
         'govuk-summary-list--no-border': noBorder,
+        [additionalClassName || '']: additionalClassName,
       })}
     >
       {Children.toArray(children).filter(child =>
