+++ src/explore-education-statistics-common/src/components/SummaryListItem.tsx
@@ -1,16 +1,31 @@
 import React, { ReactNode } from 'react';
+import classNames from 'classnames';
 
 interface Props {
   actions?: ReactNode;
   children?: ReactNode;
   term: string;
+  detailsNoMargin?: boolean;
 }
 
-const SummaryListItem = ({ actions, children, term }: Props) => {
+const SummaryListItem = ({
+  actions,
+  children,
+  term,
+  detailsNoMargin,
+}: Props) => {
   return (
     <div className="govuk-summary-list__row">
       <dt className="govuk-summary-list__key">{term}</dt>
-      {children && <dd className="govuk-summary-list__value">{children}</dd>}
+      {children && (
+        <dd
+          className={classNames('govuk-summary-list__value', {
+            'dfe-details-no-margin': detailsNoMargin,
+          })}
+        >
+          {children}
+        </dd>
+      )}
       {actions && <dd className="govuk-summary-list__actions">{actions}</dd>}
     </div>
   );
